+++ polars/polars-core/src/frame/cross_join.rs
@@ -118,7 +118,7 @@ impl DataFrame {
     ) -> PolarsResult<DataFrame> {
         let (l_df, r_df) = self.cross_join_dfs(other, slice, true)?;
 
-        self.finish_join(l_df, r_df, suffix)
+        _finish_join(l_df, r_df, suffix)
     }
 }
