+++ polars/polars-lazy/polars-pipe/Cargo.toml
@@ -19,6 +19,7 @@ polars-utils = { version = "0.25.1", path = "../../polars-utils" }
 rayon.workspace = true
 
 [features]
+compile = []
 csv-file = ["polars-plan/csv-file", "polars-io/csv-file"]
 parquet = ["polars-plan/parquet", "polars-io/parquet"]
 nightly = ["polars-core/nightly", "polars-utils/nightly", "hashbrown/nightly"]
