+++ polars/polars-lazy/polars-pipe/src/lib.rs
@@ -1,8 +1,12 @@
+#[cfg(feature = "compile")]
 mod executors;
+#[cfg(feature = "compile")]
 pub mod expressions;
+#[cfg(feature = "compile")]
 pub mod operators;
+#[cfg(feature = "compile")]
 pub mod pipeline;
 
 // ideal chunk size we strive to
-#[cfg(any(feature = "cross_join", feature = "csv-file"))]
+#[cfg(feature = "compile")]
 pub(crate) const CHUNK_SIZE: usize = 50_000;
