+++ polars/polars-lazy/polars-plan/Cargo.toml
@@ -85,7 +85,7 @@ ewma = ["polars-core/ewma"]
 dot_diagram = []
 unique_counts = ["polars-core/unique_counts"]
 log = ["polars-ops/log"]
-chunked_ids = []
+chunked_ids = ["polars-core/chunked_ids"]
 list_to_struct = ["polars-ops/list_to_struct"]
 # python = ["pyo3"]
 row_hash = ["polars-core/row_hash", "polars-ops/hash"]
