+++ polars/polars-lazy/src/physical_plan/streaming/mod.rs
@@ -1,4 +1,3 @@
 mod convert;
 
-#[cfg(any(feature = "csv-file", feature = "parquet"))]
 pub(crate) use convert::insert_streaming_nodes;
