+++ polars/src/lib.rs
@@ -169,6 +169,7 @@
 //!     - `lazy_regex` - Use regexes in [column selection](crate::lazy::dsl::col)
 //!     - `dot_diagram` - Create dot diagrams from lazy logical plans.
 //! * `sql` - Pass SQL queries to polars.
+//! * `streaming` - Be able to process datasets that are larger than RAM.
 //! * `random` - Generate arrays with randomly sampled values
 //! * `ndarray`- Convert from `DataFrame` to `ndarray`
 //! * `temporal` - Conversions between [Chrono](https://docs.rs/chrono/) and Polars for temporal data types
