+++ py-polars/Cargo.toml
@@ -50,6 +50,7 @@ pct_change = ["polars/pct_change"]
 repeat_by = ["polars/repeat_by"]
 # also includes simd
 nightly = ["polars/nightly"]
+streaming = ["polars/streaming"]
 meta = ["polars/meta"]
 search_sorted = ["polars/search_sorted"]
 decompress = ["polars/decompress"]
@@ -103,6 +104,7 @@ all = [
 default = [
   "all",
   "nightly",
+  "streaming",
 ]
 
 [dependencies.polars]
